+++ examples/src/examples/graphics/normals-and-tangents/config.mjs
@@ -1,0 +1,6 @@
+/**
+ * @type {import('../../../../types.mjs').ExampleConfig}
+ */
+export default {
+    WEBGPU_ENABLED: true,
+};
